refactor: moved logging_config + logging_filters to seperate logging module
