+++ backend/main.py
@@ -2,15 +2,14 @@ from fastapi import FastAPI
 from backend.routes.test import router as test_router
 from backend.routes.upload import router as upload_router
 from backend.routes.chat import router as chat_router
-from backend.core.config import Settings
-from backend.core.logging_config import setup_logging
+from backend.logging.logging_config import setup_logging
 from backend.middleware.request_id import register_request_id_middleware
 import logging
-
 
 app = FastAPI()
 
 logger = logging.getLogger(__name__)
+
 setup_logging()
 
 # request id middleware
